+++ backend/canisters/local_user_index/impl/src/updates/c2c_notify_events.rs
@@ -2,9 +2,19 @@ use crate::guards::caller_is_user_index_canister;
 use crate::{mutate_state, RuntimeState};
 use canister_api_macros::update_msgpack;
 use canister_tracing_macros::trace;
-use local_user_index_canister::c2c_notify_user_index_events::{Args, Event, Response};
+use local_user_index_canister::c2c_notify_events::{Response::*, *};
+use local_user_index_canister::Event;
 use tracing::info;
-use types::{PhoneNumberConfirmed, ReferredUserRegistered, StorageUpgraded, UserEvent, UserSuspended, UsernameChanged};
+use user_canister::{
+    Event as UserEvent, PhoneNumberConfirmed, ReferredUserRegistered, StorageUpgraded, UserJoinedGroup, UserSuspended,
+    UsernameChanged,
+};
+
+#[update_msgpack(guard = "caller_is_user_index_canister")]
+#[trace]
+fn c2c_notify_events(args: Args) -> Response {
+    mutate_state(|state| c2c_notify_user_index_events_impl(args, state))
+}
 
 #[update_msgpack(guard = "caller_is_user_index_canister")]
 #[trace]
@@ -16,8 +26,7 @@ fn c2c_notify_user_index_events_impl(args: Args, runtime_state: &mut RuntimeStat
     for event in args.events {
         handle_event(event, runtime_state);
     }
-
-    Response::Success
+    Success
 }
 
 fn handle_event(event: Event, runtime_state: &mut RuntimeState) {
@@ -82,10 +91,14 @@ fn handle_event(event: Event, runtime_state: &mut RuntimeState) {
             info!("Max concurrent canister upgrades set to {}", ev.value);
         }
         Event::UserJoinedGroup(ev) => {
-            runtime_state
-                .data
-                .user_event_sync_queue
-                .push(ev.user_id.into(), UserEvent::UserJoinedGroup(Box::new(ev)));
+            runtime_state.data.user_event_sync_queue.push(
+                ev.user_id.into(),
+                UserEvent::UserJoinedGroup(Box::new(UserJoinedGroup {
+                    chat_id: ev.chat_id,
+                    as_super_admin: ev.as_super_admin,
+                    latest_message_index: ev.latest_message_index,
+                })),
+            );
         }
     }
     crate::jobs::sync_events_to_user_canisters::start_job_if_required(runtime_state);